star_modify

+++ src/main/java/com/start/deli_home/Question/Controller/QuestionController.java
@@ -1,6 +1,8 @@
 package com.start.deli_home.Question.Controller;
 
 
+import com.start.deli_home.Member.Entity.Member;
+import com.start.deli_home.Member.Service.MemberService;
 import com.start.deli_home.Question.Entity.Question;
 import com.start.deli_home.Question.Service.QuestionService;
 import com.start.deli_home.Question.QuestionForm.QuestionForm;
@@ -13,12 +15,16 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.ui.Model;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -27,6 +33,7 @@ import java.util.List;
 public class QuestionController {
 
     private final QuestionService questionService;
+    private final MemberService memberService;
 
     @GetMapping("/list")
     public String questionList(Model model,
@@ -52,44 +59,61 @@ public class QuestionController {
         model.addAttribute("question", question);
         return "question_detail";
     }
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/create")
     public String questionCreate(QuestionForm questionForm){
         return "question_form";
     }
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create")
     public String questionCreate(@Valid QuestionForm questionForm, BindingResult bindingResult,
-                                 @RequestParam("images") List<MultipartFile> images){
+                                 @RequestParam("images") List<MultipartFile> images, Principal principal){
         if (bindingResult.hasErrors()){
             return "question_form";
         }
+        Member member = this.memberService.getMember(principal.getName());
+
         this.questionService.create(questionForm.getSubject(), questionForm.getContent(),questionForm.getCategory(),images,
                 questionForm.getAddress(),questionForm.getIntroduce(),questionForm.getTime(),questionForm.getMenu(),
-                questionForm.getPhone(),questionForm.getShopName(),questionForm.getFoodType());
+                questionForm.getPhone(),questionForm.getShopName(),questionForm.getFoodType(),member);
         return "redirect:/question/list";
     }
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String questionModify(QuestionForm questionForm, @PathVariable("id") Integer id) {
+    public String questionModify(QuestionForm questionForm, @PathVariable("id") Integer id,Principal principal) {
         Question question = this.questionService.getQuestion(id);
+
+        if (!question.getAuthor().getUsername().equals(principal.getName())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
 
         questionForm.setSubject(question.getSubject());
         questionForm.setContent(question.getContent());
         return "question_form";
     }
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{id}")
     public String questionModify(@Valid QuestionForm questionForm, BindingResult bindingResult,
-                                 @PathVariable("id") Integer id) {
+                                 @PathVariable("id") Integer id,Principal principal) {
         if (bindingResult.hasErrors()) {
             return "question_form";
         }
         Question question = this.questionService.getQuestion(id);
+        if (!question.getAuthor().getUsername().equals(principal.getName())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
 
         this.questionService.modify(question, questionForm.getSubject(), questionForm.getContent());
         return String.format("redirect:/question/detail/%s", id);
     }
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
     public String questionDelete(@Valid QuestionForm questionForm, BindingResult bindingResult,
-                                 @PathVariable("id")Integer id){
+                                 @PathVariable("id")Integer id,Principal principal){
         Question question = this.questionService.getQuestion(id);
+        if (!question.getAuthor().getUsername().equals(principal.getName())){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
         this.questionService.delete(question);
         return "redirect:/question/list";
     }
